% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FODRDataset.R
\name{fodr_dataset}
\alias{fodr_dataset}
\title{initialize a dataset}
\usage{
fodr_dataset(portal, id)
}
\arguments{
\item{portal}{a character in \code{\link{list_portals}}}

\item{id}{a character}
}
\description{
A wrapper around \code{FODRDataset$new(portal, id)} for convenience.
}
\examples{
\dontrun{
votes <- fodr_dataset("paris", "resultats-des-votes-budget-participatif-2016")
votes
}

}

