% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FODRPortal.R
\name{fodr_portal}
\alias{fodr_portal}
\title{initialize a portal}
\usage{
fodr_portal(portal)
}
\arguments{
\item{portal}{a character}
}
\description{
A wrapper around \code{FODRPortal$new(portal)} for convenience.
}
\examples{
\dontrun{
portal <- fodr_portal("paris")
portal
}

}
