% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FODRDataset.R
\docType{class}
\name{FODRDataset}
\alias{FODRDataset}
\title{Main dataset class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FODRDataset
}
\arguments{
\item{portal}{a character, must be one of the available portals}

\item{id}{a character, the dataset identifier}
}
\value{
An object of class \code{\link{FODRDataset}} with methods designed to retrieve data from an open dataset.
}
\description{
This is the entry point to retrieve records from a dataset. Initialize a \code{FODRDataset}
with a \code{portal} and an \code{id}.
}
\section{Fields}{

\describe{
\item{\code{portal}}{a character, must be one of the available portals}

\item{\code{data}}{a data.frame returned by the \code{\link[=dataset_get]{get}} method}

\item{\code{fields}}{a character vector}

\item{\code{facets}}{a character vector or variables that can be used to filter results}

\item{\code{id}}{a character, the dataset id}

\item{\code{info}}{a list of five elements:
\itemize{
 \item features: a character vector of available services on this dataset, \emph{unused}
 \item metas: a list of meta-information about the dataset (publisher, language, theme, etc.)
 \item attachments: a list of downloadable files related to this dataset, if any, \emph{not downloadable from R}
 \item alternative_exports: \emph{unknown purpose}
 \item billing_plans: \emph{unknown purpose}
}}

\item{\code{sortables}}{a character vector containing a subset of \strong{fields}, indicates on which fields sorting is allowed}

\item{\code{url}}{a character, the actual url sent to the API}
}}
\section{Methods}{

\describe{
  \item{\code{\link{get_attachments}}}{This method retrieves attachments from the dataset.}
  \item{\code{\link{get_records}}}{This method retrieves records from the dataset.}}
}
\keyword{datasets}

