% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FODRProvider.R
\docType{data}
\name{FODRProvider}
\alias{FODRProvider}
\title{Main provider class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FODRProvider
}
\description{
Main provider class
}
\keyword{datasets}

