% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFDataset.R
\docType{data}
\name{ODFDataset}
\alias{ODFDataset}
\title{Main dataset class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ODFDataset
}
\description{
Main dataset class
}
\keyword{datasets}

