% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFDomain.R
\docType{data}
\name{ODFDomain}
\alias{ODFDomain}
\title{Main domain class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ODFDomain
}
\description{
Main domain class
}
\keyword{datasets}

