% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodr.R
\name{fodr}
\alias{fodr}
\title{fodr: fetch French Open Data with R}
\description{
Fetch data from various French Open Data portals. Use 
\code{\link{fodr_dataset}} and \code{\link{fodr_portal}} to retrieve records.
}

