% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FODRPortal.R
\docType{class}
\name{FODRPortal}
\alias{FODRPortal}
\title{main portal class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\value{
An object of class \code{\link{FODRPortal}} with methods designed to retrieve datasets from an open data portal.
}
\description{
This is the entry point to retrieve datasets from a portal. Initialize a \code{FODRPortal} with a \code{portal} using the \code{\link{fodr_portal}} wrapper.
}
\section{Fields}{

\describe{
\item{\code{portal}}{a character, the portal name}

\item{\code{data}}{a list of \code{\link{FODRDataset}} objects}

\item{\code{facets}}{a character vector of variables that can be used to filter results. For a 
\code{\link{FODRPortal}}, these are constant.}

\item{\code{sortables}}{a character vector, indicates on which fields sorting is allowed. For a 
\code{\link{FODRPortal}}, these are constant.}

\item{\code{themes}}{a character vector, the unique themes datasets on the portals can be associated with}
}}
\section{Methods}{

\describe{
  \item{\code{\link[=portal_search]{search}}}{This method retrieves datasets from the portal.}}
}
\keyword{datasets}

