% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrappers}
\alias{fodr_dataset}
\alias{fodr_portal}
\alias{wrappers}
\title{Wrappers}
\usage{
fodr_portal(portal)

fodr_dataset(portal, id)
}
\arguments{
\item{portal}{a character}

\item{id}{a character}
}
\description{
These functions wrap the class initializers.
}

