% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrappers}
\alias{fodr_dataset}
\alias{fodr_provider}
\alias{wrappers}
\title{Wrappers}
\usage{
fodr_provider(provider)

fodr_dataset(provider, id)
}
\arguments{
\item{provider}{a character}

\item{id}{a character}
}
\description{
These functions wrap the class initializers.
}

